# 1_design/api-contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Model API 監控系統
  description: |
    Model API 即時監控系統提供完整的 API 監控、指標查詢、告警管理和實時數據推送功能。
    
    ## 主要功能
    - 🏥 **系統健康監控** - 服務狀態檢查和依賴監控
    - 📊 **指標查詢** - 實時和歷史指標數據查詢
    - 🚨 **告警管理** - 智能告警規則和通知
    - 📈 **儀表板數據** - 可視化數據支持
    - ⚡ **實時推送** - WebSocket 實時數據流
    
    ## 認證方式
    
    支持兩種 API Key 認證方式：
    
    1. **Header 認證** (推薦)
       ```
       X-API-Key: your-api-key
       ```
    
    2. **Bearer Token 認證**
       ```
       Authorization: Bearer your-api-key
       ```
       
    ## 限制
    - API 限制: 300 requests/minute
    - WebSocket 連接: 5 concurrent connections
    - 數據保留: 30 天
  version: "1.0.0"
  contact:
    name: API 支援
    email: support@monitor.example.com
    url: https://docs.monitor.example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://monitor.example.com/terms
servers:
  - url: http://localhost:8001
    description: 開發環境
  - url: https://test-api.monitor.example.com
    description: 測試環境
  - url: https://api.monitor.example.com
    description: 生產環境
security:
  - ApiKeyAuth: []
  - BearerAuth: []
paths:
...
  /v1/dashboards/overview:
    get:
      summary: 儀表板概覽
      description: 獲取儀表板總覽數據
      operationId: getDashboardOverview
      tags: ["儀表板"]
      parameters:
        - name: time_range
          in: query
          description: 時間範圍
          schema:
            type: string
            enum: [1h, 6h, 24h, 7d, 30d]
            default: 1h
      responses:
        "200":
          description: 儀表板概覽數據
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardOverviewResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
...
  /v1/dashboards/realtime:
    get:
      summary: 實時儀表板
      description: 獲取實時儀表板數據
      operationId: getRealtimeDashboard
      tags: ["儀表板"]
      responses:
        "200":
          description: 實時儀表板數據
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RealtimeDashboardResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API Key 認證
    BearerAuth:
      type: http
      scheme: bearer
      description: Bearer Token 認證
  schemas:
...
    ErrorResponse:
      type: object
      required: [success, error, timestamp]
      properties:
        success:
          type: boolean
          example: false
        error:
          $ref: "#/components/schemas/ErrorDetail"
        timestamp:
          type: string
          format: date-time
    ErrorDetail:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: "INVALID_API_KEY"
        message:
          type: string
          example: "無效的 API Key"
        developer_message:
          type: string
          example: "請檢查您的 API Key 是否正確"
...
    DashboardOverviewResponse:
      type: object
      required: [success, data, timestamp]
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: "#/components/schemas/DashboardOverviewData"
        timestamp:
          type: string
          format: date-time
    DashboardOverviewData:
      type: object
      required: [system_overview, service_health, alerts_summary, top_services]
      properties:
        system_overview:
          $ref: "#/components/schemas/SystemOverview"
        service_health:
          $ref: "#/components/schemas/ServiceHealthOverview"
        alerts_summary:
          $ref: "#/components/schemas/AlertsSummaryOverview"
        top_services:
          type: array
          items:
            $ref: "#/components/schemas/TopService"
    SystemOverview:
      type: object
      required: [total_services, total_endpoints, total_requests, total_errors, avg_qps, avg_error_rate, avg_response_time]
      properties:
        total_services:
          type: integer
        total_endpoints:
          type: integer
        total_requests:
          type: integer
        total_errors:
          type: integer
        avg_qps:
          type: number
          format: float
        avg_error_rate:
          type: number
          format: float
        avg_response_time:
          type: number
          format: float
    ServiceHealthOverview:
      type: object
      required: [healthy, degraded, unhealthy, unknown, total, health_percentage]
      properties:
        healthy:
          type: integer
        degraded:
          type: integer
        unhealthy:
          type: integer
        unknown:
          type: integer
        total:
          type: integer
        health_percentage:
          type: integer
          minimum: 0
          maximum: 100
    AlertsSummaryOverview:
      type: object
      required: [total_active, critical, high, medium, low]
      properties:
        total_active:
          type: integer
        critical:
          type: integer
        high:
          type: integer
        medium:
          type: integer
        low:
          type: integer
    TopService:
      type: object
      required: [service_name, qps, error_rate, avg_response_time]
      properties:
        service_name:
          type: string
        qps:
          type: number
          format: float
        error_rate:
          type: number
          format: float
        avg_response_time:
          type: number
          format: float
...
    RealtimeDashboardResponse:
      type: object
      required: [success, data, timestamp]
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: "#/components/schemas/RealtimeDashboardData"
        timestamp:
          type: string
          format: date-time
    RealtimeDashboardData:
      type: object
      required: [real_time_metrics, services_status, alerts_summary, last_updated]
      properties:
        real_time_metrics:
          $ref: "#/components/schemas/RealtimeMetricsOverview"
        services_status:
          type: array
          items:
            $ref: "#/components/schemas/ServiceStatus"
        alerts_summary:
          $ref: "#/components/schemas/AlertsSummaryOverview"
        last_updated:
          type: string
          format: date-time
    RealtimeMetricsOverview:
      type: object
      required: [total_qps, overall_error_rate, avg_response_time, total_services]
      properties:
        total_qps:
          type: number
          format: float
        overall_error_rate:
          type: number
          format: float
        avg_response_time:
          type: number
          format: float
        total_services:
          type: integer
    ServiceStatus:
      type: object
      required: [service_name, status, qps, error_rate]
      properties:
        service_name:
          type: string
        status:
          type: string
          enum: [healthy, degraded, unhealthy, unknown]
        qps:
          type: number
          format: float
        error_rate:
          type: number
          format: float
  responses:
...
    UnauthorizedError:
      description: 認證失敗
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            error:
              code: "INVALID_API_KEY"
              message: "無效的 API Key"
              developer_message: "請檢查您的 API Key 是否正確"
            timestamp: "2025-07-01T10:30:00Z"
...